Adicionando controllers e services

+++ projeto-se/src/main/java/com/example/projetose/entities/Paciente.java
@@ -34,7 +34,7 @@ public class Paciente {
     private Double altura;
 
     @Column
-    private Double peso;
+    private Integer peso;
 
     @OneToMany(cascade = CascadeType.ALL)
     private List<Consulta> consultas;
